+++ .env
@@ -1,1 +1,1 @@
-DATABASE_URL=postgres://insi:default@localhost/substrate_archive
+DATABASE_URL=postgres://archive:default@localhost/substrate_archive
