+++ polkadot-archive/Cargo.toml
@@ -13,6 +13,13 @@ codec = { package = "parity-scale-codec", version = "1.0.0" }
 env_logger = "0.7.0"
 substrate-archive = { path = "../" }
 
+[dependencies.polkadot-primitives]
+git = "https://github.com/insipx/polkadot"
+package = "polkadot-primitives"
+default-features = false
+branch = "insipx/archive-patches"
+features = ["std"]
+
 [dependencies.polkadot-runtime]
 git = "https://github.com/insipx/polkadot"
 package = "polkadot-runtime"
