+++ src/archive.rs
@@ -50,20 +50,19 @@ pub struct Archive<T: System> {
 impl<T> Archive<T> where T: System {
 
     pub fn new() -> Result<Self, ArchiveError> {
+        crate::util::init_logger(log::LevelFilter::Error); // TODO move this into polkadot-archive
         let mut runtime = Runtime::new()?;
         let rpc = runtime.block_on(Rpc::<T>::new(url::Url::parse("ws://127.0.0.1:9944")?))?;
         let db = Database::new()?;
         let (rpc, db) = (Arc::new(rpc), Arc::new(db));
-        match runtime.block_on(rpc.metadata(None)) {
-            Ok(v) => println!("METADATA: {:?}", v.1),
-            Err(e) => error!("Failed to receive metadata {:?}", e)
-        };
+        debug!("METADATA: {}", rpc.metadata());
+        debug!("KEYS: {:?}", rpc.keys());
         Ok( Self { rpc, db, runtime })
     }
 
     pub fn run(mut self) -> Result<(), ArchiveError> {
         let (sender, receiver) = mpsc::unbounded();
-        crate::util::init_logger(log::LevelFilter::Error); // TODO move this into polkadot-archive
+        // self.runtime.block_on(self.rpc.clone().all_storage(sender.clone()))?;
         self.runtime.spawn(
             self.rpc
                 .clone()
