+++ src/rpc.rs
@@ -16,9 +16,8 @@
 
 mod substrate_rpc;
 use self::substrate_rpc::SubstrateRpc;
-use runtime_metadata::RuntimeMetadataPrefixed;
 use log::{debug, warn, error, trace};
-use futures::{Future, Stream, sync::mpsc::UnboundedSender, future::join_all};
+use futures::{Future, Stream, sync::mpsc::UnboundedSender, future::{self, join_all}};
 use runtime_primitives::traits::Header as HeaderTrait;
 use substrate_primitives::{storage::StorageKey, twox_128};
 use substrate_rpc_primitives::number::NumberOrHex;
@@ -41,7 +40,18 @@ use crate::{
 pub struct Rpc<T: System> {
     _marker: PhantomData<T>,
     url: url::Url,
-    keys: Vec<StorageKey>
+    keys: Vec<StorageKey>,
+    metadata: Metadata
+}
+
+impl<T> Rpc<T> where T: System {
+    pub fn metadata(&self) -> &Metadata {
+        &self.metadata
+    }
+
+    pub fn keys(&self) -> &Vec<StorageKey> {
+        &self.keys
+    }
 }
 
 impl<T> Rpc<T> where T: System {
@@ -64,12 +74,12 @@ impl<T> Rpc<T> where T: System {
             })
     }
 
-    /// Query all the storage for a block
-    pub fn all_storage(self: Arc<Self>, hash: T::Hash, sender: UnboundedSender<Data<T>>
+    /// get all the storage for a single block
+    pub fn all_storage(self: Arc<Self>, /*hash: T::Hash,*/ sender: UnboundedSender<Data<T>>
     ) -> impl Future<Item = (), Error = ArchiveError>
     {
-
-
+        println!("USED KEYS: {:?}", self.keys);
+        future::ok(())
     }
 
     pub fn block_and_timestamp(self: Arc<Self>, hash: T::Hash, sender: UnboundedSender<Data<T>>
@@ -104,10 +114,12 @@ impl<T> Rpc<T> where T: System {
         SubstrateRpc::connect(&url)
             .and_then(|client: SubstrateRpc<T>| {
                 client.storage_keys(StorageKey(Vec::new()), None)
+                      .join(client.metadata(None))
             })
-            .map(|keys| {
+            .map(|(keys, metadata)| {
+                let keys = metadata.keys(keys);
                 Self {
-                    url, keys,
+                    url, keys, metadata,
                     _marker: PhantomData
                 }
             })
@@ -163,13 +175,18 @@ impl<T> Rpc<T> where T: System {
             })
     }
      */
-
-    pub fn metadata(&self, hash: Option<T::Hash>) -> impl Future<Item = RuntimeMetadataPrefixed, Error = ArchiveError> {
+/*
+    pub fn refresh_metadata(&mut self, hash: Option<T::Hash>) -> impl Future<Item = (), Error = ArchiveError> {
         SubstrateRpc::connect(&self.url)
             .and_then(move |client: SubstrateRpc<T>| {
                 client.metadata(hash)
+                      .and_then(|m| {
+                          self.metadata = m;
+                          future::ok(())
+                      })
             })
     }
+*/
 
     // TODO: make "Key" and "from" vectors
     // TODO: Merge 'from' and 'key' via a macro_derive on StorageKeyType, to auto-generate storage keys
